Module sui_sdk::apis

Structs

CoinReadApi

Coin Read API provides the functionality needed to get information from the Sui network regarding the coins owned by an address.

EventApi

Event API provides the functionality to fetch, query, or subscribe to events on the Sui network.

GovernanceApi

Governance API provides the staking functionality.

QuorumDriverApi

Quorum API that provides functionality to execute a transaction block and submit it to the fullnode(s).

ReadApi

The main read API structure with functions for retrieving data about different objects and transactions