Enum sui_sdk::types::ExecutionError

pub enum ExecutionError {

    InsufficientGas,
    InvalidGasObject,
    InvariantViolation,
    FeatureNotYetSupported,
    ObjectTooBig {
        object_size: u64,
        max_object_size: u64,
    },
    PackageTooBig {
        object_size: u64,
        max_object_size: u64,
    },
    CircularObjectOwnership {
        object: ObjectId,
    },
    InsufficientCoinBalance,
    CoinBalanceOverflow,
    PublishErrorNonZeroAddress,
    SuiMoveVerificationError,
    MovePrimitiveRuntimeError {
        location: Option<MoveLocation>,
    },
    MoveAbort {
        location: MoveLocation,
        code: u64,
    },
    VmVerificationOrDeserializationError,
    VmInvariantViolation,
    FunctionNotFound,
    ArityMismatch,
    TypeArityMismatch,
    NonEntryFunctionInvoked,
    CommandArgumentError {
        argument: u16,
        kind: CommandArgumentError,
    },
    TypeArgumentError {
        type_argument: u16,
        kind: TypeArgumentError,
    },
    UnusedValueWithoutDrop {
        result: u16,
        subresult: u16,
    },
    InvalidPublicFunctionReturnType {
        index: u16,
    },
    InvalidTransferObject,
    EffectsTooLarge {
        current_size: u64,
        max_size: u64,
    },
    PublishUpgradeMissingDependency,
    PublishUpgradeDependencyDowngrade,
    PackageUpgradeError {
        kind: PackageUpgradeError,
    },
    WrittenObjectsTooLarge {
        object_size: u64,
        max_object_size: u64,
    },
    CertificateDenied,
    SuiMoveVerificationTimedout,
    SharedObjectOperationNotAllowed,
    InputObjectDeleted,

}

Variants

InsufficientGas

Insufficient Gas

InvalidGasObject

Invalid Gas Object.

InvariantViolation

Invariant Violation

FeatureNotYetSupported

Attempted to used feature that is not supported yet

ObjectTooBig

Move object is larger than the maximum allowed size

Fields

object_size: u64

max_object_size: u64

PackageTooBig

Package is larger than the maximum allowed size

Fields

object_size: u64

max_object_size: u64

CircularObjectOwnership

Circular Object Ownership

Fields

object: ObjectId

InsufficientCoinBalance

Insufficient coin balance for requested operation

CoinBalanceOverflow

Coin balance overflowed an u64

PublishErrorNonZeroAddress

Publish Error, Non-zero Address. The modules in the package must have their self-addresses set to zero.

SuiMoveVerificationError

Sui Move Bytecode Verification Error.

MovePrimitiveRuntimeError

Error from a non-abort instruction. Possible causes: Arithmetic error, stack overflow, max value depth, etc.“

Fields

location: Option<MoveLocation>

MoveAbort

Move runtime abort

Fields

location: MoveLocation

code: u64

VmVerificationOrDeserializationError

Bytecode verification error.

VmInvariantViolation

MoveVm invariant violation

FunctionNotFound

Function not found

ArityMismatch

Arity mismatch for Move function. The number of arguments does not match the number of parameters

TypeArityMismatch

Type arity mismatch for Move function. Mismatch between the number of actual versus expected type arguments.

NonEntryFunctionInvoked

Non Entry Function Invoked. Move Call must start with an entry function.

CommandArgumentError

Invalid command argument

Fields

argument: u16

kind: CommandArgumentError

TypeArgumentError

Type argument error

Fields

type_argument: u16

kind: TypeArgumentError

UnusedValueWithoutDrop

Unused result without the drop ability.

Fields

result: u16

subresult: u16

InvalidPublicFunctionReturnType

Invalid public Move function signature. Unsupported return type for return value

Fields

index: u16

InvalidTransferObject

Invalid Transfer Object, object does not have public transfer.

EffectsTooLarge

Effects from the transaction are too large

Fields

current_size: u64

max_size: u64

PublishUpgradeMissingDependency

Publish or Upgrade is missing dependency

PublishUpgradeDependencyDowngrade

Publish or Upgrade dependency downgrade.

Indirect (transitive) dependency of published or upgraded package has been assigned an on-chain version that is less than the version required by one of the package’s transitive dependencies.

PackageUpgradeError

Invalid package upgrade

Fields

kind: PackageUpgradeError

WrittenObjectsTooLarge

Indicates the transaction tried to write objects too large to storage

Fields

object_size: u64

max_object_size: u64

CertificateDenied

Certificate is on the deny list

SuiMoveVerificationTimedout

Sui Move Bytecode verification timed out.

SharedObjectOperationNotAllowed

The requested shared object operation is not allowed

InputObjectDeleted

Requested shared object has been deleted