Struct sui_sdk::wallet_context::WalletContext

pub struct WalletContext {
    pub config: PersistedConfig<SuiClientConfig>,
    /* private fields */
}

Fields
config: PersistedConfig<SuiClientConfig>
Implementations

impl WalletContext

pub async fn new( config_path: &Path, request_timeout: Option<Duration>, max_concurrent_requests: Option<u64> ) -> Result<Self, Error>

pub fn get_addresses(&self) -> Vec<SuiAddress>

pub async fn get_client(&self) -> Result<SuiClient, Error>

pub fn active_address(&mut self) -> Result<SuiAddress, Error>

pub async fn get_object_ref( &self, object_id: ObjectID ) -> Result<ObjectRef, Error>

Get the latest object reference given a object id

pub async fn gas_objects( &self, address: SuiAddress ) -> Result<Vec<(u64, SuiObjectData)>, Error>

Get all the gas objects (and conveniently, gas amounts) for the address

pub async fn get_object_owner(&self, id: &ObjectID) -> Result<SuiAddress, Error>

pub async fn try_get_object_owner( &self, id: &Option<ObjectID> ) -> Result<Option<SuiAddress>, Error>

pub async fn gas_for_owner_budget( &self, address: SuiAddress, budget: u64, forbidden_gas_objects: BTreeSet<ObjectID> ) -> Result<(u64, SuiObjectData), Error>

Find a gas object which fits the budget

pub async fn get_all_gas_objects_owned_by_address( &self, address: SuiAddress ) -> Result<Vec<ObjectRef>>

pub async fn get_gas_objects_owned_by_address( &self, address: SuiAddress, limit: Option<usize> ) -> Result<Vec<ObjectRef>>

pub async fn get_one_gas_object_owned_by_address( &self, address: SuiAddress ) -> Result<Option<ObjectRef>>

Given an address, return one gas object owned by this address. The actual implementation just returns the first one returned by the read api.

pub async fn get_one_account(&self) -> Result<(SuiAddress, Vec<ObjectRef>)>

Returns one address and all gas objects owned by that address.

pub async fn get_one_gas_object( &self ) -> Result<Option<(SuiAddress, ObjectRef)>>

Return a gas object owned by an arbitrary address managed by the wallet.

pub async fn get_all_accounts_and_gas_objects( &self ) -> Result<Vec<(SuiAddress, Vec<ObjectRef>)>>

Returns all the account addresses managed by the wallet and their owned gas objects.

pub async fn get_reference_gas_price(&self) -> Result<u64, Error>

pub fn sign_transaction(&self, data: &TransactionData) -> VerifiedTransaction

Sign a transaction with a key currently managed by the WalletContext

pub async fn execute_transaction_block( &self, tx: VerifiedTransaction ) -> Result<SuiTransactionBlockResponse>

pub async fn batch_make_transfer_transactions( &self, max_txn_num: usize ) -> Vec<VerifiedTransaction>

A helper function to make Transactions with controlled accounts in WalletContext. Particularly, the wallet needs to own gas objects for transactions. However, if this function is called multiple times without any “sync” actions on gas object management, txns may fail and objects may be locked.

The param is called max_txn_num because it does not always return the exact same amount of Transactions, for example when there are not enough gas objects controlled by the WalletContext. Caller should rely on the return value to check the count.

pub async fn make_transfer_sui_transaction( &self, recipient: Option<SuiAddress>, amount: Option<u64> ) -> VerifiedTransaction

pub async fn make_staking_transaction( &self, validator_address: SuiAddress ) -> VerifiedTransaction

pub async fn make_publish_transaction( &self, path: PathBuf ) -> VerifiedTransaction

pub async fn make_publish_transaction_with_deps( &self, path: PathBuf ) -> VerifiedTransaction

pub async fn publish_basics_package(&self) -> ObjectRef

Executes a transaction to publish the basics package and returns the package object ref.

pub async fn publish_basics_package_and_make_counter( &self ) -> (ObjectRef, ObjectRef)

Executes a transaction to publish the basics package and another one to create a counter. Returns the package object ref and the counter object ref.

pub async fn increment_counter( &self, sender: SuiAddress, gas_object_id: Option<ObjectID>, package_id: ObjectID, counter_id: ObjectID, initial_shared_version: SequenceNumber ) -> SuiTransactionBlockResponse

Executes a transaction to increment a counter object. Must be called after calling publish_basics_package_and_make_counter.

pub async fn publish_nfts_package( &self ) -> (ObjectID, ObjectID, TransactionDigest)

Executes a transaction to publish the nfts package and returns the package id, id of the gas object used, and the digest of the transaction.

pub async fn create_devnet_nft( &self, package_id: ObjectID ) -> (SuiAddress, ObjectID, TransactionDigest)

Pre-requisite: publish_nfts_package must be called before this function. Executes a transaction to create an NFT and returns the sender address, the object id of the NFT, and the digest of the transaction.

pub async fn delete_devnet_nft( &self, sender: SuiAddress, package_id: ObjectID, nft_to_delete: ObjectRef ) -> SuiTransactionBlockResponse

Executes a transaction to delete the given NFT.